requirements.txt
Django==3.2.5
django-parler==2.2